Add stbt.press(hold_secs) parameter, and stbt.pressing context manager
Allows holding down a key on the remote-control for a certain duration,
or (the stbt.pressing context manager) while a block of code executes.

Only implemented for LIRC and Roku controls. I've tested that LIRC will
be sending the repeat signal if so configured in the lircd.conf. Note
that you might want to increase lircd's "--repeat-max" (it defaults to
600 repeats, i.e. 18 seconds for a 30ms signal including gap). LIRC
implements this by scheduling a SIGALARM for the gap duration, and when
woken up will send one more signal and decrement its "repeat_max"
counter.

I haven't implemented the other controls. I specifically know that:

- The IRNetBox doesn't support separate keydown/keyup commands.
- The CEC specification says that the receiver should assume a keyup
  event if it hasn't seen one within 550ms or so. For press-and-hold the
  initiator is supposed to send repeated keydown commands 450ms apart,
  followed by the final keyup. (CEC 13.13.3)