HdmiCecControl: Implement press-and-hold
CEC spec section 13.13.3 says that the receiver should assume a keyup if
it hasn't seen one within a receiver-determined timeframe which can't be
less than 550ms. For press-and-hold the initiator should send repeated
keydown commands (200 to 450ms apart) followed by the final keyup.

So in `keydown()` we start a thread that sends keydown commands 450ms
apart (don't know if we need more sophisticated timing -- according to
the spec the receiver is supposed to receive them this distance apart,
so technically we should sleep for 450ms minus however long the transmit
took). We don't raise an exception for commands that don't get
acked (after the first one) because we've seen that the FireTV needs
interpress_delay_secs=5 just when using single presses.