tests: Fix check for printing unicode exceptions
Commit 9809814 (stbt-run: Fix printing unicode tracebacks) added the
`assert_correct_unicode_error` helper function for the self-tests.
This helper function checks "mylog" which is written to by the first
`stbt run` invocation, but not the second one.

Here we're using `unbuffer` to simulate interactive use. Presumably
that's still just as effective as it was, even now that we're
redirecting the output to a file, because our test-runner (run-tests.sh)
was already redirecting the test's output to a file called "log".