stbt.ocr: Add text_color_threshold parameter
The selftest screenshot is a menu where the highlighted (selected) item
is white text, and the other menu items are blue text. Without
specifying text_color, ocr reads both items correctly. We can use
text_color to find the current selection, but with the current threshold
of 25 it reads "GUIde" instead of "Guide". The thresholded image is
completely missing the dot on top of the "i". Doubling the threshold to
50 fixes the mis-read (and the thresholded image looks much nicer).
Maybe we should change the default threshold -- we need more testcases.