Prep: Add private `logging.draw_on` function
Allows us to draw on a frame in the sink pipeline even if all we have is
a reference to a frame.  It does this by storing a weak reference to the
`SinkPipeline` with every frame.

This will allow us to better decouple our video-processing functions from
`DeviceUnderTest`.

Currently this is all private because `SinkPipeline` itself is private and
storing a `weakref` on every `Frame` is a bit icky anyway.  I do think it
makes conceptual sense that if we have a `time` member on `Frame` we should
also have some indication of where the frame was grabbed from.  For a
public API this could be an `int` or a `string` or something, but given
that it's not public for now a `weakref` is simpler.