Move `wait_for_motion` and `detect_motion` into `motion.py`
...and add a `frames` parameter to both.

They are no longer members of `DeviceUnderTest` so we no-longer need to
have forwarding implementations in `stbt/__init__.py`.

I think this general model makes some sense.  The different image
processing bits are no longer bound so tightly together.  But we make a
concession to convenience - frames defaults to `stbt.frames()`.  By
introducing this impurity we maintain the convenience of the existing API.
It makes the tests cleaner too.