stbt.frames: Change return type to Iterator[Frame]
...from Iterator[(Frame, int)].

This is a breaking change, but the error will happen immediately and the
error message will be somewhat obvious: "ValueError: too many values to
unpack". Also, there should be very few uses of `stbt.frames()` in any
given test-pack.

The second field of the yielded tuple was a timestamp in
nanoseconds (GStreamer format). This has been deprecated since we added
`stbt.Frame.time` in v26, 2 years ago.

Now seems like a good time to do this because we've just added a
`frames` parameter to `detect_motion` and `wait_for_motion`. This API is
much more convenient.