Refactor: Move match functions into own file
As a result `detect_match` and `wait_for_match` gained a `frames`
parameter.

This is analogous to fdca78fdf and 21f4440b1 but for the match functions,
which took up a significant proportion of core.py.

`press_until_match` is still in `_stbt.core` because it involves pressing
as well.