stbt-lint: stbt-frame-object-missing-frame checker: Also check class constructors
Inside a FrameObject property, if you do `stbt.match("image.png")`
instead of `stbt.match("image.png", frame=self._frame)`, stbt-lint will
warn you. However it wasn't warning if you were instantiating a class
that takes a `frame` parameter. This is a common use-case because
sometimes we want to return a FrameObject as the result of a different
FrameObject's property. For example, a FrameObject called `Dialog`
might have a property called `button` that returns an instance of a
`Button` FrameObject class.