stbt lint: Report images that exist but aren't committed to git
This is a common & annoying problem -- stbt-lint passes locally but
fails on your CI server (as do the actual tests!).

I've used `subprocess` instead of `dulwich` because the latter is hard
to use (the version on Ubuntu 16.04 doesn't have
`dulwich.porcelain.ls_files`) and its Apache 2.0 license isn't
compatible with LGPL v2.1.