stbt lint: Report full path to image
It used to say "images/reference.png" (that is, a path relative to the
python file). Now, if you're running stbt-lint from the root of your
test-pack, it will say "tests/images/reference.png" -- the path relative
to your current directory. This makes it easier for you to copy and
paste into a "git add" command. It also shows you exactly what path
stbt-lint resolved the image lookup to, which might help debugging any
issues with said lookup.