Python API: Add `stbt.OcrPatterns.IP_ADDRESS`
The OCR system that stb-tester uses was trained against long-form text so
it expects to read normal words.  This can confuse it when it comes across
text that aren't words - like IP addresses.

In these instances you can use `tesseract_user_patterns` to tell tesseract
to expect a string of a certain form, but it can be unclear what to pass to
it.  Now you can write:

    stbt.ocr(tesseract_user_patterns=stbt.OcrPatterns.IP_ADDRESS)

This tells tesseract to expect something of the form `x.x.x.x` where x is
a number fixing the misread.