stbt.match: Add `structure_weighting`
We use Sobel edge detection to improve the matching.  This is then
taken into account when constructing the heatmap.  This more closely
matches human expectations at the cost of being slower to run.

The weighting is done by raising the values to a power before
multiplying them together.  This is inspired by SSIM where seperate
brightness, contrast and structure terms are combined with the same
technique.