FrameObject API: Add `FrameObject.refresh`
This makes navigation functions easier to write.  You can pass in a
`FrameObject` instance rather than passing in a `FrameObject` class
and then reinstantiating each time.  It also provides an extra point
of customisation for the `FrameObject` implementer.  For example you
could have a `FrameObject` instance parameterised in some way, and
you want to propogate that parameterisation through to new frames.
This allows the `FrameObject` implementer to do that.