Add initial configuration for CircleCI
Unlike Travis, CircleCI makes it easy to run tests inside a docker
container. So we'll build docker images for Ubuntu 16.04 and Ubuntu
18.04 (the latter in a future commit), push them to Docker
Hub (manually, needs to be done by the Stb-tester project owners but
only when we need to install new dependencies), and then CircleCI will
pull the image from Docker Hub each time it runs the tests.