ocr: test_that_ligatures...: Handle OCR mis-read
with tesseract 4.00 and the legacy engine data from [tessdata] commit
590567f.

Note that "sillyness" is mis-spelled in the source image. But
"ligatures" is a real word so you'd expect OCR to read it correctly.

Tesseract's new LSTM model reads this correctly (but it has other
regressions evident in our other unit tests).

[tessdata]: https://github.com/tesseract-ocr/tessdata