circleci: Disable parallelism of integrationtests
This is to fix the CI job hanging on CircleCI, eventually failing with
"Too long with no output (exceeded 10m0s)".

It does take a lot longer: 8-9 minutes (for the whole job), compared to
a job from yesterday that took 3 minutes. But maybe the CirceCI job
executors are more heavily loaded today -- after all we hadn't seen the
tests hang before.