stbt.OcrMode: Add new page segmentation modes
SPARSE_TEXT and SPARSE_TEXT_OSD were added in 3.03 (Ubuntu 14.04). We
don't officially support older versions than that.

RAW_LINE was added in 3.04 (Ubuntu 16.04).

None of these are documented in the man page but here's the output from
`tesseract --help-extra`:

> 11    Sparse text. Find as much text as possible in no particular order.
> 12    Sparse text with OSD. [Orientation and Script Detection]
> 13    Raw line. Treat the image as a single text line,
>       bypassing hacks that are Tesseract-specific.

I think "Raw line" is designed for passing a line of text directly to
the LSTM neural net, which takes an input 36px high and variable width.