circleci: Upgrade pylint, but test stbt-lint with system pylint
We use pylint for 2 things:

1. To lint stb-tester's own code (`make check-pylint`).
2. To lint user's code (stbt-lint, which is a custom pylint plugin).

On both CI environments (Ubuntu 16.04 and 18.04) we need the same
version of pylint to keep maintenance cost down for `make check-pylint`.
For example, the older pylint on Ubuntu 16.04 (v1.5.2) doesn't
understand that `Enum`s are unsubscriptable.

So on both CI environments (Ubuntu 16.04 and 18.04) I have upgraded:

| Package | New version | Old version |
|---------+-------------+-------------|
| pylint  |       1.8.3 |       1.6.4 |
| astroid |       1.6.0 |       1.4.8 |
| isort   |       4.3.4 |       3.9.0 |

The new versions are the same that Ubuntu 18.04 provides, so on 18.04 we
just use the system packages.

However for stbt-lint we do want to test that stbt-lint works with the
system-provided pylint.
On 18.04 that's easy, now that we're using the system-provided pylint
for `make check-pylint`, that's easy.
On 16.04, I have added an extra step that uninstalls the pip packages
and runs the stbt-lint tests against the system pylint.