pylint: Disable len-as-condition, no-else-return, ungrouped-imports
len-as-condition complains about `if len(x) == 0`. I like it because
it's explicit.

`no-else-return` wants us to write this | instead of this:

    if x:              | if x:
        return y       |     return y
    return z           | else
                       |     return z

Personally I prefer pylint's version on the left, but @wmanley likes the
version on the right and there are too many to fix right now.

`ungrouped-imports` complains about this:

    from gi.repository import GLib, GObject, Gst

I think it wants them grouped with parentheses: (GLib, GObject, Gst).
I don't think it matters so it's easier to disable than to fix them all.