API: Add return value from `stbt.press`
A common pattern for benchmarking is:

    stbt.press('KEY_XXX')
    start_time = time.time()
    x = wait_until(xyz)
    duration = x.time - start_time

This moved the `time.time()` bits into `stbt.press` itself where it
can do a better job as it is `interpress_delay_secs` aware.

This was motivated by some changes I want to make to the transition
APIs.  When writing page objects you have to deal with transitions
between different pages.  For example a menu might know that it has
to press `'KEY_OK'` to launch the EPG, but only the EPG knows how
to wait until the EPG has finished loading.  In this case the `key`
argument to `press_and_wait` is defined by the `Menu` page object,
whereas the `mask` would be defined by `EPG`. This API is a step
towards being able to implement this as:

    r = stbt.press('KEY_OK')
    page = EPG.wait_until_loaded(r)

This motivates the `frame_before` member.