stbt.match: Add new MatchMethod SQDIFF
The sum-of-squared-differences is normalised by the number of pixels in
the sum. The existing SQDIFF_NORMED (where the normalisation is done by
OpenCV) normalises by the *intensities* of the pixels, so it exaggerates
differences for dark images (and it doesn't work at all for completely
black images, as it essentially divides by zero).