stbt.match: Refactor: heatmap is always "darkest is best"
In `_match_template` we change the heatmap so that it can always be
interpreted the same way, regardless of the MatchMethod that was used to
generate it. This greatly simplifies the caller of `_match_template`.

Before, SQDIFF or SQDIFF_NORMED meant that you needed to find the
darkest pixel in the heatmap; and CCORR_NORMED or CCOEFF_NORMED the
brightest pixel. I chose the "darkest is best" representation for all
cases, to avoid unnecessary calculations for SQDIFF/SQDIFF_NORMED. I
don't really care about CCORR_NORMED/CCOEFF_NORMED -- they were never
particularly useful and we're only keeping them for backwards
compatibility.