tests: Add tests with transparent reference image
I have committed the generated png images because I spent ages trying to
debug test failures not realising that these images are generated. You
can still re-generate them with `make tests/buttons.png` but they are
not automatically re-generated from `make check`.

Note that we have to generate them using `rsvg-convert`, not Inkscape.
If I try to open the SVG in Inkscape, it changes the size of the image
because I have a new version of Inkscape (0.92.3 from Ubuntu 18.04).