stbt.pressing: Move `sys.exc_info()` into `except` block
`sys.exc_info()` usually returns `(None, None, None)` if an exception
isn't currently being handled, but this isn't always the case.
[PEP 343] says: "`sys.exc_info()` has very complex semantics and it is
perfectly possible that it returns the exception information for an
exception that was caught ages ago."  We have seen this happen in
cases involving complex nested context managers.

Now we only call `sys.exc_info()` if we *know* that we are handling an
exception.

[PEP 343]: https://www.python.org/dev/peps/pep-0343/