test_auto_selftest_caching: Disable stbt.match performance optimisation
With the new SQDIFF default threshold of 0.98, the pyramid optimisation
means that a smaller part of the video-frame is considered at the larger
sizes of the pyramid, so `stbt.match` runs much faster. For example, on
my laptop:

    ==> without_cache.time <==
    4.43
    ==> cold_cache.time <==
    3.72
    ==> hot_cache.time <==
    3.22

With the caching it's still faster, but not twice as fast.

By disabling the "pyramid" performance optimisation, the times are
closer to 6.9s / 4.9s / 3.2s.