Fix test_template_annotation_labels
We're looking for stb-tester's annotation that says:

> match('videotestsrc-checkers-8.png')

...but the reference image was matching towards the right-hand side of
the video, so the text was truncated. The text we were trying to
find (with `wait_for_match`) wasn't entirely there.

I suppose the new match method (SQDIFF) must be matching at a different
location than the old method. Note that both locations are equally valid
because the reference image genuinely wasn't present in the video frame.

I have changed the test so that it looks for something that *is*
present, and that's guaranteed to match far enough to the left that the
full label will be visible.

Note that `pattern=ball` always starts from the same position in the
middle of the frame.

Note: I have also added `--save-screenshot never` to test.py so that
it doesn't overwrite the screenshot from verify.py (test.py is killed
so it would log a screenshot at that point).