Added support for RedRat X
Made some small modifications to irnetbox.py making it possible to
use a RedRat X IR blaster with stbt. This basically just involved
adding the RRX PID (13) to class NetBoxTypes and modifying the if/else
in method irsend_raw.

This will treat the RRX same as an IRNetBox allowing for communication
with the device using ethernet. Please note that it will NOT support
the RRX's USB interface.

I have tested this implementation thoroughly on my side and find that
it works well. According to the people at RedRat chances are good it
will also work with IRNetBox 4 but I have been unable to verify this.