stbt.match: Merge neighbouring ROIs from previous pyramid level
If the previous pyramid level (that is, running `cv2.matchTemplate` on a
smaller version of the image) identified many Regions Of Interest (ROIs)
that are small and close together, then at the next pyramid level we
have to invoke `cv2.matchTemplate` (now on the full-sized image)
multiple times.

This is still a worthwhile optimisation compared to running
`cv2.matchTemplate` on the entire full-sized image; but if we merge
ROIs that are very close together, it's even faster.

This effect is particularly pronounced, now that we've tightened the
match threshold, if you're looking for an image of a word and the frame
has a full paragraph of text; the word might (tentatively) match at
every word in the paragraph, but not at the spaces between the words.

In the above scenario, I measured 47ms vs. 69ms, on my laptop.

N.B. The morphological operation "Closing" means to dilate and then
erode -- it has the effect of closing holes or gaps. Using a square
kernel (all ones) is 4 times faster than using an elliptical kernel
(250µs for the `c2v.morphologyEx` vs. 1.1ms).