stbt.match: Performance optimisation: Don't threshold last matchTemplate heatmap
Here we create `roi_mask` which is used in the next iteration of the
`for` loop. In the last iteration we don't need to create a new
`roi_mask`, because it isn't used after the `for` loop.

This saves up to 1ms on my laptop.

With this, `test_that_cache_speeds_up_match_all` started failing:

    >       assert uncached_time > (cached_time * 4)
    E       assert 0.014809131622314453 > (0.003849029541015625 * 4)