sqdiff: Add numba implementation for comparison purposes
Our C implementation usually outperforms the numba implementation by a
factor of 4-8x, but sometimes it is slower than numba.  This seems to be
caused by fixed overheads as it's only ever faster on very small templates
when the fixed-costs dominate.  For example:

    All times in ms             numpy		 numba
    type        numpy	numba	C	speedup	speedup	size	alignment

    template    0.03	0.01	0.02	1.40	0.53	7 x 480 	5
    with mask	0.05	0.01	0.02	2.53	0.31	7 x 480 	5
    unmasked    0.04	0.01	0.02	2.27	0.62	7 x 480 	5

In this instance the comparisons are taking ~10-20us.

On larger templates (the case we care about) the C implementation is
faster:

    All times in ms                     numpy   numba
    type    	numpy	numba	C	speedup	speedup	size	alignment

    template 	28.59	2.74	0.27	106.06	10.15	1280 x 720 	0
    with mask	34.48	1.04	0.63	54.68	1.64	1280 x 720 	0
    unmasked 	31.47	2.04	0.61	51.55	3.33	1280 x 720 	0

In this case the C implementation is taking ~270us without transparency
and 620us with.

Interestingly the numba implementation is faster with transparency than
without - but it's still slower than C.