stbt.match: Invert the meaning of confirm_threshold
Now 0 means "less strict" and 1.0 means "completely strict". i.e. it's
measuring the similarity, rather than the difference (error).

This is for consistency with the `match_threshold`. The discrepancy has
caused some confusion in the past.

In all the user test-scripts that we've seen, confirm_threshold is never
changed from its default values, except to set it to an absurdly high
value (where the intent was really to disable the confirm pass; you can
do that with `confirm_method=ConfirmMethod.NONE`).