stbt.match: Optimization: Merge regions-of-interest that overlap completely
Currently we're getting rois like this:

![](https://user-images.githubusercontent.com/12424/52785764-b7c26c80-3050-11e9-84ff-acc46480e2ab.png)

Note that `cv2.findContours` with `cv2.RETR_EXTERNAL` is working
correctly: those internal regions aren't internal contours, they're
completely outside the large contour. It's only when we take the
*bounding rectangle* of the large contour, that the smaller ones fall
inside it. So we have to manually filter them out. I couldn't find a
suitable OpenCV function to do it so I've implemented a simple quadratic
loop.

After `_merge_regions` it looks like this:

![](https://user-images.githubusercontent.com/12424/52785909-230c3e80-3051-11e9-93d0-1e41f0cd5c14.png)

This is a simple quadratic implementation, but it shouldn't matter
because we don't expect huge numbers of regions. On my laptop it takes:

| number of regions |  ms |
|-------------------|-----|
|                20 | 0.2 |
|               200 |   4 |
|              2000 |  27 |