ImageLogger: Fix size & position of region annotations
`max-width:100%` [of the parent container] is necessary so that portrait
screenshots don't exceed the width of the parent div.

`display:inline-block` is necessary so that the parent container is the
width of its contents (i.e. of the image), for portrait
screenshots (e.g. from a phone) that are narrower than the page.

Note that the region div's position is specified as a percentage of the
container div (`.annotated_image`).