All python scripts: Change shebangs to /usr/bin/python
...in preparation for Python 3 support (we plan to support Python 2 & 3
from the same codebase).

As far as I can tell, you set this to /usr/bin/python in your sources,
and `setuptools` modifies the shebang line when it installs the files.
We don't use setuptools because we have so many non-python dependencies
but it seems to be the (a?) standard[1].

[1]: https://packaging.python.org/guides/tool-recommendations/#packaging-tool-recommendations