Python3: Delay converting filename to bytes
On Python 2 `cv2.imread` fails if given unicode, so `stbt.load_image`
converts the given string to bytes.

But on Python 3 `stbt.load_image` was failing to `os.path.join` a
unicode string with a byte string.

To support both we still give `cv2.imread` a bytes string, but we
convert it to bytes at the last possible moment.