Python3: test_that_matchresult_str...: Handle Python 2 & 3 formatting
I give it | Python 2 formats it as | Python 3 formats it as |
----------|------------------------|------------------------|
"..." †   | u"..."                 | "..."                  |
u"..."    | u"..."                 | "..."                  |
b"..."    | "..."                  | b"..."                 |

† Because we use `from __future__ import unicode_literals` everywhere.

So `str(MatchResult)` will always give a different result on Python 2
than Python 3. We have to use a regex to explicitly ignore the `u`
prefix. This solution is similar to pytest's "ALLOW_UNICODE" which is
for doctests.