Remove `stbt batch` tool
We want to focus this project on the stbt *library*. The surrounding
tooling is becoming a maintenance burden, and for all I know it doesn't
work properly because it's hard to write tests for it. We / our
customers don't use `stbt batch`.

If you want to use this feel free to maintain it in a separate repo as
per the [license]. But I would recommend using pytest instead (see
<https://github.com/stb-tester/stb-tester/issues/297>).

[license]: https://github.com/stb-tester/stb-tester/blob/master/LICENSE