Remove `stbt auto-selftest` tool
We want to focus this project on the stbt *library*. The surrounding
tooling is becoming a maintenance burden, and despite our best efforts
this tool hasn't been adopted by anyone as far as we know. We do still
strongly believe in the value of the auto-selftest technique, but we
don't have the resources to continue to maintain it. Our customers use a
similar, but GUI-based, tool with a completely separate implementation:
The Stb-tester [Object Repository].

If you want to use `stbt auto-selftest` feel free to maintain it in a
separate repo as per the [license].

[Object Repository]: https://stb-tester.com/manual/object-repository
[license]: https://github.com/stb-tester/stb-tester/blob/master/LICENSE