stbt.match: Pyramid optimisation: Ignore pixels on edge of reference image
[cv2.pyrDown] blurs the image with a 5x5 gaussian kernel before
downscaling it. With the reference image, the edge pixels will be
blurred with the neighbouring pixels based on the [borderType]
parameter, which defaults to `cv2.BORDER_REFLECT_101`. So for a 1-row
reference image with these pixels:

       abcde

pixel "a" will be blurred with [c b a b c] (in this explanation I'm only
considering the horizontal component of the blur, for simplicity).

On the other hand the frame, if you're considering a potential match
position somewhere that's not on the edge of the frame, might look like
this:

    xyzabcdefgh

So pixel "a" will be blurred with [y z a b c]. So the frame value that
you're comparing against will be different than the reference value.

Normally this doesn't matter: there are so many pixels in the reference
image that *aren't* on the edge, that the *average* squared-difference
isn't affected much by these discrepancies. However if you have a
reference image that's mostly transparent, and the only non-transparent
pixels are near the edge, you can encounter this issue. See the selftest
images I'm committing here for an example: The reference image is trying
to find a "selection" (thin border) around a tile, so the content of the
tile is ignored (by making it transparent in the reference image).

This bug only applies to MatchMethod.SQDIFF. With the other methods, we
relax the threshold so much that we're unlikely to hit this bug.
Relaxing the threshold by that much with SQDIFF would essentially make
the pyramid optimisation a waste of time.

Note that I've also changed `borderType` to `cv2.BORDER_REPLICATE`. This
didn't help any with the failing test-case but it seems more correct.
Because we're now ignoring the edge pixels in the reference image it
won't make any difference to the reference image, but it will affect the
edges of the actual frame that we're searching in.

[cv2.pyrDown]: https://docs.opencv.org/3.2.0/d4/d86/group__imgproc__filter.html#gaf9bba239dfca11654cb7f50f889fc2ff
[borderType]: https://docs.opencv.org/3.2.0/d2/de8/group__core__array.html#ga209f2f4869e304c82d07739337eae7c5