pylint3: Ignore "Do not raise StopIteration in generator"
This is warning about the new behaviour of [PEP 479] added in Python
3.5: If a generator function receives a StopIteration exception and
doesn't handle it, the exception will be converted to RuntimeError
instead of bubbling out as a StopIteration. The correct way of
signalling the end from a generator function is simply to return.

This doesn't apply here, as `stbt.frames` will never raise
StopIteration.

[PEP 479]: https://www.python.org/dev/peps/pep-0479/