python3: VideoTestSrcControl.press: Fix conversion to bytes
The GStreamer property has to be bytes (at least on Python 2 -- haven't
tested on Python 3 yet), but we were converting the key to bytes too
early. The membership test was `if key not in [0, u"smpte", ...]` which
was working on Python 2 (if `key` was bytes it still evaluates to True)
but it won't work on Python 3.