Python3: stbt lint: Open sys.stderr in binary mode
In stbt lint we read from the `pylint` subprocess's stderr, filter out
irrelevant warnings that we can't silence otherwise (typically
originating from C libraries like GStreamer or GLib), and then write the
rest of the output to `sys.stderr`.

The input is always in bytes:
`p = subprocess.Popen(..., stderr=subprocess.PIPE); stderr.read()`
returns bytes on both Python 2 and 3.

However Python 3's `sys.stderr.write` only accepts unicode strings, by
default. So here we open `sys.stderr` in binary mode. This means that we
are treating all the input & output as bytes; we don't have to decode
the input (which we were doing explicitly) and we don't have to encode
it back before writing (which Python was doing implicitly when we called
`sys.stderr.write`).

Side-effect of this change: On Python3, the stderr output will be
buffered. Line-buffering can only be set in text mode. On Python 2 the
output is still line-buffered.