stbt lint: Set stderr to unbuffered
On Python 3, files opened in binary mode can't be set to line-buffered,
so they default to fully-buffered. We are reading & writing 1 line at a
time, and unless the lines are ridiculously long each write will be
atomic. So with this change the behaviour is --in practice-- line
buffered, which is a much better user experience when running
interactively in your terminal.