Remove "restart_source" config setting
This was a workaround for a bug in the Hauppauge HDPVR video-capture
device, which is ancient and unreliable hardware. As far as I know,
nobody uses this setting.

Note that we *do* restart the source pipeline if we receive EOS (always,
regardless of configuration). This setting also restarted the source
pipeline on underrun (no buffers received, but without an EOS either).