Python3: stbt run: Implement sane_unicode_and_exception_handling for Python 3
I'm too afraid to touch the Python 2 version, so I have provided two
separate implementations.

In spite of the name, the Python 3 version doesn't really do anything
special with regards to unicode. It's merely there to do two things:

a. Print our "FAIL" line (where stbt batch's "failure-reason" comes
   from) which has some special logic to generate an assertion message
   from the traceback if you did `assert` without an explicit message.

b. Return 1 for test failure, 2 for test error.

I have expanded our self-test to cover more cases, including a check for
the assertion logging described in (a) above; and at the same time I've
simplified it by making it more table-like.

As you can see in the test, the behaviour is the same on Python 2 & 3
except when the exception message is a byte-string: On Python 2 it's
printed normally, but on Python 3 it's printed as:

    RuntimeError: b'R\xc3\xb6thlisberger'

(I couldn't figure out the right amount of escaping to make grep find
the backslashes so I have replaced them in the "expected output" with a
regex's ".".)

This was the only place we use `python-kitchen` so we don't need it at
all for Python 3.