Python3: stbt.ocr: Allow passing a byte-string as tesseract_user_words
Because why not. If you pass a single str, we wrap it in a list so that
tesseract sees the single word, not a set of single characters as
separate "words". With this commit, you get the same behaviour if you
pass a byte-string. You're unlikely to do this on Python 3 but it
doesn't hurt to support it and it fixes
`test_user_dictionary_with_non_english_language`.