Python3: stbt.ocr: Fix debug html rendering
For some reason, python2-jinja2 and python3-jinja2 behave differently,
even though they're both the same jinja2 version (2.10).

On Python2, and undefined variable in the Jinja template is ignored, but
on Python3 it triggers an error if you try to use it. This is the
relevant part of the template in ocr.py:

    {{ annotated_image(result) }}

`imglog.data["result"]` was specified for `stbt.match_text` (a
`TextMatchResult`) but not for `stbt.ocr`.

Then in `_stbt.logging._draw_annotated_image`, the first parameter
(that this `result` variable is passed as) is called `regions`. We do
this:

    if isinstance(regions, Region):

That's fine, Jinja doesn't complain and `isinstance()` evaluates to
False. But then we do this:

    elif hasattr(regions, "region"):

...and Jinja raises "jinja2.exceptions.UndefinedError: 'result' is
undefined." But only on Python 3.

You can specify how Jinja treats undefined variables in the templates:
http://jinja.pocoo.org/docs/2.10/api/#undefined-types

> The default undefined type [...] can be printed and iterated over, but
> every other access will raise `jinja2.exceptions.UndefinedError`.

It seems that in practice you can do `isinstance` checks as well as
printing & iterating it; but you can't call `hasattr`. But on Python 2
you can. Who knows why. BTW none of Jinja's other pre-defined classes
(for changing the undefined behaviour) would have helped here, and I
don't think it's worth trying to define my own custom class to override
the behaviour. This `result` variable was the only variable that was
missing and also gets used in both `stbt.ocr` and `stbt.match_text`; all
the other `match_text`-specific variables are protected behind
`{% if match_text %} ... {% endif %}`.

This fixes `test_stbt_debug.py::test_ocr_debug`.