CircleCI: Don't run pep8 in Python2 job
This is because pep8 is the only thing that still requires Python3 (it's
implemented using Python3) after the last few commits where I've been
removing any unnecessary dependencies from the CI images.

This means that the Python2 image doesn't have Python3 installed at all,
which I believe is important to test the Python 2 vs. 3 packaging (when
we come to implement it).

We do still run pep8 in the Python3 CI job, which will catch the same
issues. It's a shame to have a difference between the 2 jobs but I think
it's worth it.