run-tests.sh: Handle Ctrl-C correctly
Previously the SIGINT would be sent to the bash subprocess that's
running the individual test. So the test would stop, but then the next
test in the list would start running.

Now we run that subprocess in the background, so that the parent
bash process also receives the SIGINT and exits immediately.