Python3: stbt.get_frames: Fix comparison of None with int
On Python2 `max(10, None)` is 10. On Python3 it's a TypeError:
'>' not supported between instances of 'NoneType' and 'int'.

N.B. This timeout of 10s is where we raise `NoVideo` if we don't get any
frames at all from the source pipeline. It was added for the HDPVR
according to the comment in commit 4947e4b ("stbt.Display.frames: Always
return at least 1 frame") but I don't know if it's also relevant for
other hardware so I don't want to remove it -- too risky. It's certainly
easier to debug if you get `NoVideo` than if the process hangs...

This fixes `test_detect_motion_changing_mask` and probably many others.