stbt.as_precondition: Fix wrong parameter to `traceback.format_exc()`
This takes a single parameter that is supposed to be an int: `limit`,
i.e. how many stack frames to print. Internally it gets the exception
information from `sys.exc_info()`. On Python 2 this worked anyway,
but on Python 3 it was failing with:

    TypeError: '>=' not supported between instances of 'MatchTimeout' and 'int'

This fixes `test_stbt_run_return_code_on_precondition_error`.