stbt.FrameObject repr: Limit float property values to 3 decimal places
Otherwise doctest output of floating-point values might vary from run to
run.

60 frames per second is 17ms per frame, so a 1ms precision should be
enough for any measurements related to video-frames. For example we have
been affected by this in a FrameObject that we use in our A/V sync
tests, that has a property showing the lag or lead time from the audio
to the corresponding video frame.

We only apply this to properties that have "time" or "_secs" in the
name, because we don't know what other use-cases people are using
floating-point values for. If you need this, at least now you have the
option of working "time" or "_secs" into your property name.

This won't help properties that are lists of floats, etc. But it's a
start.