keyboard: Handle non-deterministic transitions
For example the YouTube search keyboard looks like this:

    A B C D E F G
    H I J K L M N
    O P Q R S T U
    V W X Y Z - '
    SPACE CLEAR SEARCH

When you're at the second-to-last row (V W X Y Z) pressing down always
takes you to SPACE (even from Z). And then pressing up always takes you
back to the key that you were on before. This is true even if you move
around on the bottom row: If you start at W, down (to SPACE) - right (to
CLEAR) - right (to SEARCH) - up (goes back to W).

To handle this we treat the state machine as a "non-deterministic finite
automaton" (NFA): In a given state, for a given input, it might end up
in any of several possible states. In reality the behaviour of the
keyboard *is* deterministic, but only if you keep track of the previous
state that you had come from; our implementation is less complicated if
we treat it as a NFA.

So for state transitions that are "non-deterministic" in this sense, we
check (using OCR via `self.page.selection`) what state we actually
landed in, and then re-calculate the remaining path.