keyboard: Fix non-deterministic transitions
The bottom 2 rows of the YouTube search keyboard looks like this:

    V W X Y Z - '
    SPACE CLEAR SEARCH

From any node in the "V W X..." row, pressing down and up lands you back
at the starting node. Even pressing down, then right to CLEAR or SEARCH,
then back up, always lands you at the starting node. To model this in
the test-pack would require a lot of extra states, or two separate state
machines, or sub-states; so to keep the API simple we model this as a
"nondeterministic" state machine where we don't know exactly which node
we'll land on after pressing up from SPACE; we only know which are the
possible nodes.

Unfortunately this means that `networkx.shortest_path` thinks it can
take a shortcut via the SPACE node: W -down-> SPACE -up-> Z. In reality,
when it presses up from SPACE it lands back at W, not at Z. To prevent
this we add a large weight to all nondeterministic edges.

Note that the weight defaults to 1 for edges without an explicit
"weight" attribute.